Рефакторит утилитарные функции + распределяет их в соответствующие тематике модули
