+++ src/utils/render.js
@@ -1,0 +1,39 @@
+import Abstract from "../view/abstract.js";
+
+export const RenderPosition = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+export const render = (container, child, position = `beforeend`) => {
+  container = container instanceof Abstract ? container.getElement() : container;
+  child = child instanceof Abstract ? child.getElement() : child;
+
+  switch (position) {
+    case RenderPosition.AFTERBEGIN:
+      container.prepend(child);
+      break;
+    case RenderPosition.BEFOREEND:
+      container.append(child);
+      break;
+  }
+};
+
+export const renderTemplate = (container, template, position = `beforeend`) => {
+  container = container instanceof Abstract ? container.getElement() : container;
+  container.insertAdjacentHTML(position, template);
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+  return newElement.firstChild || ``;
+};
+
+export const remove = (component) => {
+  if (!(component instanceof Abstract)) {
+    throw new Error(`Can remove only components`);
+  }
+  component.getElement().remove();
+  component.removeElement();
+};
